% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{make_mat}
\alias{make_mat}
\title{Make a matrix}
\usage{
make_mat(n)
}
\arguments{
\item{n}{size of matrix}
}
\value{
A n*n square matrix
}
\description{
This function initializes a n*n matrix
}
\keyword{matrix}
